requirements.txt
flask
pdf2image
pillow
img2pdf
torch
torchvision
einops